Fix: the parser now reads legal `\S\'` sequences.
In order to achieve this the fix removes the ability to read illegal
step files that don't escape \ to \\.

This was needed because otherwise the \S\' sequence would be parsed
as "backslash","S","backslash", causing an early termination when
hitting the following apostrophe.

The EncodeBackslash test has been marked `Ignored` to signify this.
This fix I assumes that reading legal files should be preferred to
illegal ones, but maybe there's a way to read both.