Sets appropriate Header.FileDescription.Description
The value is set depending on schema version

- IFC2x3: `ViewDefinition [CoordinationView_V2.0]`
- IFC4: `ViewDefinition [ReferenceView_V1.2]`
- IFC4X3_ADD2: `ViewDefinition [ReferenceView]`

This is a up from https://github.com/xBimTeam/XbimEssentials/pull/624#issuecomment-2845124738

I've opted for a minimal change solution, because the values are already
configurable with a provided parameter.

I've also added the .editorconfig to improve style coherence and
reformatted the two modified files accordingly.