Work to speed-up SaveAs for Esent models
When using HeuristicModelProvider the performance saving an esent file
to another location was slow, because it did recreate the database.
With this fix the built-in SaveAs is invoked, which performs faster
file operations.

Also I've changed the version of FluentAssertions to avoid license
issues with 8.0.0, which requires a paid license.

Minor changes to other test files to avoid conflicting xbim file names.