Esent mishandling of abstract classes
Memory model was more resilient than Esent when it came to
models incorrectly instantiating Abstract classes.

Esent models threw exceptions where the memory model would not.

This patch enriches an existing test case to query the number of
IIfcProducts from a bad file, and ensures that the Esent model behaves
the same as memoryModel.